Fix GetSubTree depth argument type
GetSubTree takes an integer for the depth, and not a size_t.
The upcoming mapper implementation needs it correct.

Tested:  No longer throws an exception with the fix.

+++ dbus.cpp
@@ -83,7 +83,7 @@ DbusPropertyMap getAllProperties(sdbusplus::bus::bus &bus,
 }
 
 DbusSubtree getSubtree(sdbusplus::bus::bus &bus, const std::string &root,
-                       size_t depth, const std::string &interface)
+                       int depth, const std::string &interface)
 {
     DbusSubtree tree;
